Mikielina-app/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.1
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
mysqlclient==2.1.1
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.4
urllib3==1.26.15
